fix: address PR #566 review — late-scene partial capture + dedup helpers + test split
Three blocking findings from the code review:

1. Late-scene partial-capture overwrite.
   `_capture_partial_acp_trajectory` early-returned whenever
   `self._trajectory` was non-empty, so a multi-scene rollout where scene 1
   succeeded but scene 2 timed out/crashed mid-stream would drop scene 2's
   partial events: the live writer briefly put `prior + scene-2 partial` on
   disk, but the final write at `_build_rollout_result` then overwrote that
   with stale `self._trajectory` (scene 1 only). Fix: capture the current
   session's delta beyond `_session_traj_count` and extend `self._trajectory`
   on every disconnect, setting `partial_trajectory=True` /
   `trajectory_source="partial_acp"` when a delta was actually appended.

2. Duplicated event-to-trajectory schema mapping.
   `_capture_session_trajectory` (destructive, end-of-run) and
   `_snapshot_session_trajectory` (non-destructive, live) had the same
   tool_call / text-event conversion inlined twice. Extracted to
   `_events_to_trajectory` so "streaming format == final format" is a
   structural invariant instead of copy/paste discipline.

3. `tests/test_capture_trajectory.py` over 1k lines.
   Moved writer / streaming / multi-scene / unknown-update / stale-tmp
   tests out into a new `tests/test_trajectory_streaming.py` (also home to
   the new `TestMultiScenePartialCaptureFix` regression for finding #1).
   `test_capture_trajectory.py` stays scoped to capture semantics.

Defensive `getattr` on `self._acp_client.session` preserves legacy
`object()` stubs used in `tests/test_rollout_branch.py`.

2379 passed, 11 skipped, no regressions.
